backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0

# Payment Gateway
razorpay==1.4.1

# ML & Computer Vision
ultralytics==8.1.0
opencv-python-headless==4.9.0.80

# Database (Optional - Supabase)
supabase==2.3.0

# Environment & Configuration
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Rate Limiting
slowapi==0.1.9

# WebSocket
websockets==12.0

# Utilities
email-validator==2.1.0
python-dateutil==2.8.2
